pyproject.toml
[project]
name = "deer-face-embed"
version = "0.1.0"
description = "Metric learning model for red deer individual identification from facial images"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "Apache-2.0"}
authors = [{name = "Jonathan Schulze Buschhoff", email = "jsbuschhoff@gmail.com"}]
keywords = ["wildlife", "metric-learning", "re-identification", "deer", "pytorch"]
classifiers = [
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Recognition",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "pandas>=2.2.2",
    "pyarrow>=17.0.0",
    "tqdm>=4.66.5",
    "opencv-python>=4.10.0.84",
    "pytorch-metric-learning>=2.6.0",
    "typer>=0.12.5",
    "timm>=1.0.9",
    "matplotlib>=3.9.2",
    "seaborn>=0.13.2",
    "faiss-cpu>=1.9.0",
    "mlflow>=2.19.0,<3",
    "torch-summary>=1.4.5",
    "loguru>=0.7.3",
    "pydantic>=2.0",
    "pandas-stubs>=2.2.3.250308",
    "wildlife-tools>=0.0.2",
    "pycocotools>=2.0.8",
    "wildlife-datasets>=1.0.5",
    "hydra-zen>=0.14.0",
    "kaggle>=1.6.0",
    "psutil>=7.2.2",
]

[project.optional-dependencies]
cpu = [
  "torch>=2.9.1",
  "torchvision>=0.24.1",
]
cu126 = [
  "torch>=2.9.1",
  "torchvision>=0.24.1",
]
cu128 = [
  "torch>=2.9.1",
  "torchvision>=0.24.1",
]

[project.urls]
Homepage = "https://github.com/rotwild-id/deer-face-embed"
Repository = "https://github.com/rotwild-id/deer-face-embed"
Issues = "https://github.com/rotwild-id/deer-face-embed/issues"

[project.scripts]
deer-face-embed = "deer_face_embed:main"

[tool.uv]
package = true
conflicts = [
  [
    { extra = "cpu" },
    { extra = "cu126" },
    { extra = "cu128" },
  ],
]

[tool.uv.sources]
torch = [
  { index = "pytorch-cpu", extra = "cpu" },
  { index = "pytorch-cu126", extra = "cu126" },
  { index = "pytorch-cu128", extra = "cu128" },
]
torchvision = [
  { index = "pytorch-cpu", extra = "cpu" },
  { index = "pytorch-cu126", extra = "cu126" },
  { index = "pytorch-cu128", extra = "cu128" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "PTH", "B"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
docstring-code-format = true
docstring-code-line-length = 88

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "pytest-cov>=5.0.0",
    "ruff>=0.11.2",
    "ty>=0.0.1a6",
]
